feat: 내가 생성한 리뷰들 조회를 커서 기반 페이징으로 구현

+++ src/main/java/com/example/jini_umc10th/domain/review/dto/ReviewReqDTO.java
@@ -1,9 +1,11 @@
 package com.example.jini_umc10th.domain.review.dto;
 
+import java.math.BigDecimal;
+
 public class ReviewReqDTO {
 
     public record postReviewReqDTO(
-            int rating,
+            BigDecimal rating,
             String content,
             String[] imageUrl
     ) {}
